Rearrange uC pin assignment for robustness
Lots of the pins on the uC are '5V tolerant'. What this actually means
is that they don't have any clamping diodes to the positive rail and are
therefore more fragile; they cannot handle any injected current. I've
rearranged the IO assignment so that almost all of the outputs are on
the 5V tolerant pins and inputs that require clamping are on non-5V
tolerant pins. The enable lines to the output switches do not require
clamping diodes, as the discrete pull-down resistor effectively forms a
potential divider looking into the uC. The only place which requires
external clamping diodes is the UART data lines.