Rearchitect power switch
The power switch now only switches power to the 3V3 SMPS and UVLO. The
added quiescent current from the 5V SMPS when the board is switched off
is negligible. This massively reduces the amount of current flowing
through the power switch.

Also move the fan and its FET onto the supply schematic page.