Change the 3.3V SMPS and UVLO solution
The previous solution used the same SMPS IC as the 5V rail, which was
good in terms of reusing the same part, but suboptimal due to it being
designed for large currents (~3A). It also required a separate IC to
perform the UVLO functionality, which was quite expensive.

This solution uses a single IC that is much more suited to low current
draw (it is very efficient down to essentially 0A) and has a built in
micro-power comparator which can be used for UVLO. It's also a much
cheaper solution.