Connect second power switch pole to microcontroller
There is the slight possibility that the board might not power down when
the power switch is turned off. This could happen if somehow current was
being injected into the 3.3V rail from either a 12V output or the 5V
output. It should be impossible as the only place where current could be
injected is the external start button connector and this has a 4k7
resistor in series, limiting the current to 2mA max. The power LED on
the 3.3V rail draws way more current than this.

However, the second pole of the power switch is available and connecting
it to the microcontroller is free. It will allow the microcontroller to
read the state of the internal power switch for the brief time when the
SMPS is still managing to provide power.