Increase copper clearance around positive battery terminal
These terminals are push-fit and are just blocks of metal. There's very
little to prevent them from shorting to copper directly underneath. The
terminal is slightly raised in the middle, but could short to the top
copper layer around the pins. Increase the clearance to the copper
around pins of the positive battery terminal.