Pull-down 5V enable line
The original idea was that the 5V rail would be enabled by the 3V3 rail
and optionally disabled by the micro. What actually happens is that the
5V rail blips on when the board is turned off. It's just must safer to
have a pull-down on the enable line and let the micro turn the 5V rail
on. The only downside to this is that, when the micro is rebooted after
performing a firmware upgrade, we need the 5V rail to remain up. This
is achieved with a cap, which will hold the enable line high for a short
period.