Reduce height of case by 2mm
This allows the power and start buttons to protrude above the top of the
case even when they're pressed. Unfortunately this now means that the
camcon plugs will foul the top panel of the case. To prevent this there
needs to be cut-outs above each 7.5mm and 5mm camcon socket.